+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@ Release Notes
 
 0.6.0 (in progress)
 -------------------
+- replaced shared timer with instance per smartcron to allow parallel executions of smartcrons (each timer is backed by a single thread)
 - ...
 
 0.5.0